+++ api/client/banner/index.js
@@ -1,0 +1,3 @@
+const router = module.exports = require('express').Router();
+
+router.get('/', require('./get.fetch-banner-list'));