+++ config/development.js
@@ -1,14 +1,26 @@
 module.exports = {
   alidayu: {
-    appkey: '23580200',
-    appsecret: 'adb3a077dd4819ada46710a2ee6333b5',
+    appkey: '24380690',
+    appsecret: 'd296b0da0e681ee22b90eaa3cfb3107b',
     restUrl: 'http://gw.api.taobao.com/router/rest',
     smsNumSend: {
-      signName: '正辰',
-      templateCode: 'SMS_67140688',
-      maxCountPerDay: 10,
-      maxCountPerHour: 5,
-      perMsgLiveCycle: 60 * 10000
+      register: {
+        signName: '阿里大于测试专用',
+        templateCode: 'SMS_71260267',
+        maxCountPerDay: 10,
+        maxCountPerHour: 5,
+        perMsgLiveCycle: 60 * 10000,
+        product: 'Fashion World'
+      },
+
+      resetPassword: {
+        signName: '阿里大于测试专用',
+        templateCode: 'SMS_71260265',
+        maxCountPerDay: 10,
+        maxCountPerHour: 5,
+        perMsgLiveCycle: 60 * 10000,
+        product: 'Fashion World'
+      },
     }
   },