+++ models/Banner.js
@@ -8,13 +8,14 @@ const { Schema } = mongoose;
 const { Mixed, ObjectId } = Schema.Types;
 
 let schema = new Schema({
-  channelId: { type: ObjectId, default: null }
-  title: { type: String, required: true, minlength: 2, maxlength: 32, default: '' },
+  channelId: { type: ObjectId, default: null },
+  title: { type: String, required: true, minlength: 2, maxlength: 64, default: '' },
   type: { type: String, required: true, minlength: 2, maxlength: 64, default: 'url' },
   value: { type: Mixed, required: true, default: '' },
-  cover: { type: String, default: '' },
+  cover: { type: String, default: null },
   priority: { type: Number, default: 0 },
-  createAt: { type: Date, default: Date.now }
+  createAt: { type: Date, default: Date.now },
+  removeAt: { type: Date, default: null }
 });
 
 schema.virtual('coverUrl').get(function() {